requirements.txt
selenium
webdriver-manager
selenium-wire
blinker==1.7.0
beautifulsoup4
pandas
openpyxl
langchain
openai
langchain-community
langchain-openai
pydantic-settings
lxml
fastapi
uvicorn
google-cloud-storage
setuptools
python-dateutil